YJTableView
宏掉动画时间 缩短动画时间  调整 进度条颜色 进度显示不在显示小数点后的数字

+++ Lib/YJView/LYRoundProgressView.m
@@ -73,7 +73,7 @@
 {
     _progress=value>self.minValue?(value>self.maxValue?self.maxValue:value):self.minValue;
     [proStr release];
-    proStr=[[NSString alloc] initWithFormat:@"%.1f%@",_progress*100,@"%"];
+    proStr=[[NSString alloc] initWithFormat:@"%d%@",(int)(_progress*100),@"%"];
     [self setNeedsDisplay];
 }
 #pragma mark--绘图操作
